kie-remote-client: compilation fix (generics + TaskServiceClientCommandObject)

+++ kie-remote/kie-remote-client/src/main/java/org/kie/services/client/api/command/TaskServiceClientCommandObject.java
@@ -403,7 +403,7 @@ public class TaskServiceClientCommandObject extends AbstractRemoteCommandObject 
         cmd.setJaxbTask(genTask);
         JaxbStringObjectPairArray values = convertMapToJaxbStringObjectPairArray(params);
         cmd.setParameter(values);
-        return executeCommand(cmd);
+        return (Long) executeCommand(cmd);
     }
 
     @Override
